General refactor to make the codebase cleaner
