+++ README.md
@@ -4,8 +4,7 @@ This example webapp uses [IndraDB](https://github.com/indradb/indradb) to explor
 
 ## Getting started
 
-* Make sure you have python 3 installed.
+* Make sure you have rust installed.
 * Make sure you have IndraDB installed, and that the applications are available in your `PATH`.
 * Clone the repo.
-* Get a copy of the [latest wikipedia data](https://dumps.wikimedia.org/enwiki/latest/enwiki-latest-pages-articles.xml.bz2) and download it into the root directory of the repo.
-* Run `make`. The run will take a long time, as several GB worth of data needs to be decompressed, parsed and indexed. But subsequent invocations of `make` will be snappy.
+* Run `make`. This will take a long time, as it'll (1) download the wikipedia dataset if you don't have it already, (2) decompress and parse the archive, and (3) index the content into IndraDB. But subsequent invocations of `make` will be snappy.
